separate getFromContainer from Container.js
