+++ satoshi/src/renderer/core/Container.js
@@ -1,11 +1,7 @@
 // @flow
-
-import { factory } from './factory';
-
 /**
- * this export statement is for test purpose.
- * Container class should not be imported anywhere inside the application.
- * Use `getFromContainer` function.
+ * DI Container
+ * Container class should not be used via `getFromContainer` function.
  * */
 export class Container {
   _instances: { [key: ClassName]: ClassInstance };
@@ -19,18 +15,3 @@ export class Container {
     return this._instances[SomeClass.name];
   }
 }
-
-const registerInstances = (container: Container): void => {
-  factory().forEach((instance: ClassInstance) => {
-    container.register(instance);
-  });
-};
-
-const container = new Container();
-registerInstances(container);
-
-export const getFromContainer = <T>(SomeClass: Class<T>): T => {
-  return container.get(SomeClass);
-};
-
-export default getFromContainer;
